Projets/AI Projets/Baobab_AI/backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
langchain==0.2.0
langchain-community==0.2.0
faiss-cpu
tiktoken
serpapi
google-search-results
langchain-openai

# Database dependencies
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Caching
redis==5.0.1
aioredis==2.0.1

# Additional utilities
pydantic[email]==2.5.1
python-dateutil==2.8.2
pytest==7.4.3
httpx==0.25.2